+++ app/resources/views/pages/dominion/invade.blade.php
@@ -25,7 +25,7 @@
                     </div>
                 </div>
             @else
-                <form action="{{ route('dominion.invade') }}" method="post" role="form">
+                <form action="{{ route('dominion.invade') }}" method="post" role="form" id="invade_form">
                     @csrf
 
                     <div class="box box-primary">
@@ -91,10 +91,12 @@
                                                 </span>
                                             </td>
                                             <td class="text-center">
-                                                {{ $unit->power_offense }}
+                                                <span id="unit{{ $unitSlot }}_op">
+                                                    {{ (strpos($unit->power_offense, ".") !== false) ? number_format($unit->power_offense, 1) : number_format($unit->power_offense) }}
+                                                </span>
                                                 /
-                                                <span class="text-muted">
-                                                    {{ $unit->power_defense }}
+                                                <span id="unit{{ $unitSlot }}_dp" class="text-muted">
+                                                    {{ (strpos($unit->power_defense, ".") !== false) ? number_format($unit->power_defense, 1) : number_format($unit->power_defense) }}
                                                 </span>
                                             </td>
                                             <td class="text-center">
@@ -274,9 +276,37 @@
                 templateSelection: select2Template,
             });
 
+            updateUnitStats();
+
+            $('#target_dominion').change(function (e) {
+                updateUnitStats();
+            });
+
             // please forgive me for this monstrosity
             $('input[name^=\'unit\']').change(function (e) {
-                // var input = $(this);
+                calculate();
+            });
+
+            function updateUnitStats() {
+                $.get(
+                    "{{ route('api.dominion.invasion') }}?" + $('#invade_form').serialize(), {},
+                    function(response) {
+                        if(response.result == 'success') {
+                            $.each(response.units, function(slot, stats) {
+                                // Update unit stats data attributes
+                                $('#unit\\['+slot+'\\]').data('dp', stats.dp);
+                                $('#unit\\['+slot+'\\]').data('op', stats.op);
+                                // Update unit stats display
+                                $('#unit'+slot+'_dp').text(stats.dp.toLocaleString(undefined, {maximumFractionDigits: 2}));
+                                $('#unit'+slot+'_op').text(stats.op.toLocaleString(undefined, {maximumFractionDigits: 2}));
+                            });
+                            calculate();
+                        }
+                    }
+                );
+            }
+
+            function calculate() {
                 var invasionForceOPElement = $('#invasion-force-op');
                 var invasionForceBoatsElement = $('#invasion-force-boats');
                 var homeForcesOPElement = $('#home-forces-op');
@@ -284,7 +314,6 @@
                 var homeForcesBoatsElement = $('#home-forces-boats');
                 var invadeButtonElement = $('#invade-button');
                 var allUnitInputs = $('input[name^=\'unit\']');
-                var unitSlot = parseInt($(this).data('slot'));
 
                 var invadingForceOP = 0;
                 var invadingForceDP = 0;
@@ -304,11 +333,6 @@
                 var DPNeededToLeaveAtHome; // 33% rule
                 var allowedMaxOP; // 5:4 rule
 
-                // Calculate total unit OP / DP
-                var unitStatsElement = $('#unit' + unitSlot + '_stats');
-                unitStatsElement.find('.op').text((parseInt($(this).val() || 0) * (parseFloat($(this).data('op')) * OPMultiplier)).toLocaleString(undefined, {maximumFractionDigits: 2}));
-                unitStatsElement.find('.dp').text((parseInt($(this).val() || 0) * (parseFloat($(this).data('dp')) * DPMultiplier)).toLocaleString(undefined, {maximumFractionDigits: 2}));
-
                 // Calculate invading force OP / DP
                 allUnitInputs.each(function () {
                     // var unitAmount = parseInt($(this).data('amount')); // total amount at home before invading
@@ -316,6 +340,13 @@
                     var unitDP = parseFloat($(this).data('dp'));
                     var amountToSend = parseInt($(this).val() || 0);
                     var needBoat = !!$(this).data('need-boat');
+
+                    var totalUnitOP = amountToSend * unitOP * OPMultiplier;
+                    var totalUnitDP = amountToSend * unitDP * DPMultiplier;
+                    var unitSlot = parseInt($(this).data('slot'));
+                    var unitStatsElement = $('#unit' + unitSlot + '_stats');
+                    unitStatsElement.find('.op').text(totalUnitOP.toLocaleString(undefined, {maximumFractionDigits: 2}));
+                    unitStatsElement.find('.dp').text(totalUnitDP.toLocaleString(undefined, {maximumFractionDigits: 2}));
 
                     if (amountToSend === 0) {
                         return true; // continue
@@ -400,7 +431,7 @@
                 } else {
                     invadeButtonElement.removeAttr('disabled');
                 }
-            });
+            }
         })(jQuery);
 
         function select2Template(state) {