+++ app/resources/views/pages/dominion/military.blade.php
@@ -53,13 +53,13 @@
                                                 @if ($unit->power_offense == 0)
                                                     <span class="text-muted">0</span>
                                                 @else
-                                                    {{ $unit->power_offense }}
+                                                    {{ (strpos($unit->power_offense, ".") !== false) ? number_format($unit->power_offense, 1) : number_format($unit->power_offense) }}
                                                 @endif
                                                 /
                                                 @if ($unit->power_defense == 0)
                                                     <span class="text-muted">0</span>
                                                 @else
-                                                    {{ $unit->power_defense }}
+                                                    {{ (strpos($unit->power_defense, ".") !== false) ? number_format($unit->power_defense, 1) : number_format($unit->power_defense) }}
                                                 @endif
                                             </td>
                                             <td class="text-center">