+++ src/Helpers/UnitHelper.php
@@ -59,7 +59,11 @@ class UnitHelper
             'reduce_combat_losses' => 'Reduces combat losses.',
             'immortal_except_vs_icekin' => 'Immortal (except vs Icekin).',
             'fewer_casualties_offense' => '%s%% fewer casualties on offense.',
-            'counts_as_spy' => 'Each unit count as %s of a spy.',
+            'counts_as_wizard_offense' => 'Each unit counts as %s of a wizard on offense.',
+            'counts_as_wizard_defense' => 'Each unit counts as %s of a wizard on defense.',
+            'immortal_vs_land_range' => 'Immortal when attacking %s%% of your size or greater.',
+            'counts_as_spy_offense' => 'Each unit counts as %s of a spy on offense.',
+            'counts_as_spy_defense' => 'Each unit counts as %s of a spy on defense.',
         ];
 
         foreach ($race->units as $unit) {