+++ src/Models/InfoOp.php
@@ -42,7 +42,12 @@ class InfoOp extends AbstractModel
 
     public function targetDominion()
     {
-        //
+        return $this->belongsTo(Dominion::class, 'target_dominion_id');
+    }
+
+    public function targetRealm()
+    {
+        return $this->belongsTo(Realm::class, 'target_realm_id');
     }
 
 //    public function scopeNotInvalid(Builder $query): Builder