+++ src/Services/Dominion/Actions/SpellActionService.php
@@ -161,7 +161,7 @@ class SpellActionService
                 ],
                 'redirect' =>
                     $this->spellHelper->isInfoOpSpell($spellKey) && $result['success']
-                        ? route('dominion.op-center.show', $target->id)
+                        ? $result['redirect']
                         : null,
             ] + $result;
     }
@@ -235,8 +235,8 @@ class SpellActionService
     {
         $spellInfo = $this->spellHelper->getSpellInfo($spellKey, $dominion->race);
 
-        $selfWpa = $this->militaryCalculator->getWizardRatio($dominion);
-        $targetWpa = $this->militaryCalculator->getWizardRatio($target);
+        $selfWpa = $this->militaryCalculator->getWizardRatio($dominion, 'offense');
+        $targetWpa = $this->militaryCalculator->getWizardRatio($target, 'defense');
 
         // You need at least some positive WPA to cast info ops
         if ($selfWpa === 0.0) {
@@ -277,17 +277,30 @@ class SpellActionService
 
         // todo: take Energy Mirror into account with 20% spell reflect (either show your info or give the infoop to the target)
 
-        $infoOp = InfoOp::firstOrNew([
-            'source_realm_id' => $dominion->realm->id,
-            'target_dominion_id' => $target->id,
-            'type' => $spellKey,
-        ], [
-            'source_dominion_id' => $dominion->id,
-        ]);
+        if ($spellKey === 'clairvoyance') {
+            $infoOp = InfoOp::firstOrNew([
+                'source_realm_id' => $dominion->realm->id,
+                'target_realm_id' => $target->realm->id,
+                'type' => $spellKey,
+            ], [
+                'source_dominion_id' => $dominion->id,
+                'target_dominion_id' => $target->id,
+            ]);
+        } else {
+            $infoOp = InfoOp::firstOrNew([
+                'source_realm_id' => $dominion->realm->id,
+                'target_dominion_id' => $target->id,
+                'type' => $spellKey,
+            ], [
+                'target_realm_id' => $target->realm->id,
+                'source_dominion_id' => $dominion->id,
+            ]);
+        }
 
         if ($infoOp->exists) {
             // Overwrite casted_by_dominion_id for the newer data
             $infoOp->source_dominion_id = $dominion->id;
+            $infoOp->target_dominion_id = $target->id;
         }
 
         switch ($spellKey) {
@@ -309,7 +322,10 @@ class SpellActionService
                     'resource_ore' => $target->resource_ore,
                     'resource_gems' => $target->resource_gems,
                     'resource_tech' => $target->resource_tech,
-                    'resource_boats' => $target->resource_boats + $this->queueService->getInvasionQueueTotalByResource($dominion, 'resource_boats'),
+                    'resource_boats' => $target->resource_boats + $this->queueService->getInvasionQueueTotalByResource(
+                            $dominion,
+                            'resource_boats'
+                        ),
 
                     'morale' => $target->morale,
                     'military_draftees' => $target->military_draftees,
@@ -331,11 +347,11 @@ class SpellActionService
                 $infoOp->data = $this->spellCalculator->getActiveSpells($target);
                 break;
 
-//            case 'clairvoyance':
-//                $infoOp->data = [
-            // tc
-//                ];
-//                break;
+            case 'clairvoyance':
+                $infoOp->data = [
+                    'targetRealmId' => $target->realm->id
+                ];
+                break;
 
 //            case 'disclosure':
 //                $infoOp->data = [];
@@ -349,11 +365,16 @@ class SpellActionService
         $infoOp->updated_at = now(); // todo: fixable with ->save(['touch'])?
         $infoOp->save();
 
+        $redirect = route('dominion.op-center.show', $target);
+        if ($spellKey === 'clairvoyance') {
+            $redirect = route('dominion.op-center.clairvoyance', $target->realm->id);
+        }
+
         return [
             'success' => true,
             'message' => 'Your wizards cast the spell successfully, and a wealth of information appears before you.',
             'wizardStrengthCost' => 2,
-            'redirect' => route('dominion.op-center.show', $target),
+            'redirect' => $redirect,
         ];
     }