+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 
 ## [Unreleased]
 
+## [0.6.2-1] - 2019-06-10
+### Added
+- Added Clairvoyance spell
+- Added new races: Dark Elf, Gnome, Halfling, Icekin, Sylvan, and Troll
+
+### Changed
+- Changed timestamp displays from relative server time (eg '13 hours ago') to absolute server time (eg '2019-06-19 13:33:37'). A setting will be added in the future for this, including round time (eg 'Day 12 Hour 23')
+
 ## [0.6.2] - 2019-06-05
 ### Changed
 - Changed realm size to 6 (from 12)
@@ -228,7 +236,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 ### Added
 - This CHANGELOG file.
 
-[Unreleased]: https://github.com/WaveHack/OpenDominion/compare/0.6.2...HEAD
+[Unreleased]: https://github.com/WaveHack/OpenDominion/compare/0.6.2-1...HEAD
+[0.6.2-1]: https://github.com/WaveHack/OpenDominion/compare/0.6.2...0.6.2-1
 [0.6.2]: https://github.com/WaveHack/OpenDominion/compare/0.6.1-5...0.6.2
 [0.6.1-5]: https://github.com/WaveHack/OpenDominion/compare/0.6.1-4...0.6.1-5
 [0.6.1-4]: https://github.com/WaveHack/OpenDominion/compare/0.6.1-3...0.6.1-4